🚚 [Core] move context-related utils into a separate file
